LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-buttons-activation-expected.txt



FAIL CSS Test: ::scroll-button() activation assert_equals: expected 604 but got 0
